fix: accept fractional lookbackHours (int? -> double?) so 15m picker no longer 400s
The TimeRangePicker '15m' preset sends lookbackHours=0.25, which ASP.NET
Core could not bind to int? and returned 400 Bad Request. Widening the
parameter to double? fixes the binding; TimeSpan.FromHours already accepts
double, so no other code changes are needed. Integration test added to pin
the contract.

+++ backend/src/HotFixAmbulance.Api/Program.cs
@@ -67,7 +67,7 @@ app.MapGet("/api/apis", (ApisConfig apis) =>
 
 app.MapPost("/api/triage/{apiName}", async (
     string apiName,
-    [FromQuery] int? lookbackHours,
+    [FromQuery] double? lookbackHours,
     [FromQuery] DateTimeOffset? fromUtc,
     [FromQuery] DateTimeOffset? toUtc,
     TriageService service,
